Update transaction decimals

+++ app/requests/models.py
@@ -218,11 +218,11 @@ class Transaction(models.Model):
 class transaction_description(models.Model):
     tracking_number = models.ForeignKey(Transaction, models.DO_NOTHING, to_field='tracking_number')
     provided_data = models.CharField(max_length=255, blank=True, null=True)
-    regular_price = models.FloatField()
+    regular_price = models.DecimalField(max_digits=19, decimal_places=2)
     regular_quantity = models.IntegerField()
     discount_price = models.FloatField()
     discount_quantity = models.IntegerField()
-    total = models.FloatField()
+    total = models.DecimalField(max_digits=19, decimal_places=2)
     user = models.ForeignKey(AuthUser, models.DO_NOTHING, blank=True, null=True)
 
     @property
